BE/PetPal/users/tokens.py
import datetime

import jwt

from django.conf import settings


def generate_activation_token(user):
    payload = {
        'id': str(user.id),
        'exp': datetime.datetime.utcnow() + datetime.timedelta(hours=48),
        'iat': datetime.datetime.utcnow()
    }

    return jwt.encode(payload, settings.SECRET_KEY, algorithm='HS256')
